+++ test/plots/index.js
@@ -136,8 +136,8 @@ export {default as moviesProfitByGenre} from "./movies-profit-by-genre.js";
 export {default as musicRevenue} from "./music-revenue.js";
 export {default as musicRevenueBars} from "./music-revenue-bars.js";
 export {default as musicRevenueBin} from "./music-revenue-bin.js";
-export {default as musicRevenueContrived} from "./music-revenue-contrived.js";
 export {default as musicRevenueGroup} from "./music-revenue-group.js";
+export {default as musicRevenueWiggle} from "./music-revenue-wiggle.js";
 export {default as ordinalBar} from "./ordinal-bar.js";
 export {default as penguinAnnotated} from "./penguin-annotated.js";
 export {default as penguinCulmen} from "./penguin-culmen.js";
